Realizando shows, falta outos detalhes

+++ Projeto_Java/src/br/edu/ufca/dados/Vocalista.java
@@ -2,7 +2,7 @@ package br.edu.ufca.dados;
 
 public class Vocalista extends MusicoAbstrato {
 	
-	public Vocalista(String nome,int exp,int preco) {
+	public Vocalista(String nome,double exp,int preco) {
 		super(nome,exp,preco);
 	}
 }
